requirements.txt
streamlit==1.32.2
streamlit-authenticator==0.2.3
pyyaml==6.0.1
pandas==2.1.4
numpy==1.26.4
python-dateutil==2.9.0
pytz==2024.1
tzdata==2024.1
setuptools==70.0.0
gspread==6.0.2
oauth2client==4.1.3
gspread-dataframe==3.3.1